Move deployment code to deployment-utils.js
